+++ packages/common-substrate/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add alias `parseProjectManifest`, also follow type of `INetworkCommonModule` (#2462)
+
 ## [3.8.1] - 2024-06-21
 ### Fixed
 - Fix `EventFilter` incorrectly extend `BlockFilter`, lead dictionary error (#2463)